Add files via upload
You have 3 towers and N disks of different sizes which can slide on to any tower. The puzzle starts with disks sorted in ascending order of size from top to bottom (i.e., each disk sits on top of an even larger one).You have the following constraints:(1) Only one disk can be moved at a time.(2) A disk is slid off the top of one tower onto another tower.(3) A disk cannot be placed on top of a smaller disk. Write a program to move the disks from the first tower to the last using Stacks